+++ tests/models/test_model_schemas.py
@@ -1,9 +1,10 @@
 import json
 import re
+from collections.abc import Callable
 from copy import deepcopy
 from importlib.resources import files
 from itertools import zip_longest
-from typing import Any, Callable
+from typing import Any
 
 import pytest