+++ mex/common/logging.py
@@ -1,8 +1,9 @@
 import logging
 import logging.config
+from collections.abc import Callable, Generator
 from datetime import datetime
 from functools import wraps
-from typing import Any, Callable, Generator, Optional, TypeVar, Union
+from typing import Any, TypeVar
 
 import click
 
@@ -59,12 +60,12 @@ def watch(
     return wrapper
 
 
-def get_ts(ts: Optional[datetime] = None) -> str:
+def get_ts(ts: datetime | None = None) -> str:
     """Get a styled timestamp tag for prefixing log messages."""
     return click.style(f"[{ts or datetime.now()}]", fg="bright_yellow")
 
 
-def echo(text: Union[str, bytes], ts: Optional[datetime] = None, **styles: Any) -> None:
+def echo(text: str | bytes, ts: datetime | None = None, **styles: Any) -> None:
     """Echo the given text with the given styles and the current timestamp prefix.
 
     Args: