+++ mex/common/primary_source/extract.py
@@ -1,5 +1,5 @@
 import json
-from typing import Generator
+from collections.abc import Generator
 
 from mex.common.logging import watch
 from mex.common.primary_source.models import SeedPrimarySource