+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- port `get_inner_types` from `mex-backend` to `mex.common.utils`
+
 ### Changes
+
+- modernize typing with syntactic sugar
+- simplify `BaseModel._get_list_field_names` using `get_inner_types`
 
 ### Deprecated