+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added subclasses with specific resolution YearMonth, YearMonthDay, YearMonthDayTime
 - modernize typing with syntactic sugar
 - simplify `BaseModel._get_list_field_names` using `get_inner_types`
+- switch to pdm
 
 ### Deprecated