+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-## [0.22.0] - 2024-03-18
+## [0.22.0] - 2024-03-19
 
 ### Added
 
@@ -31,6 +31,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added subclasses with specific resolution YearMonth, YearMonthDay, YearMonthDayTime
 - modernize typing with syntactic sugar
 - simplify `BaseModel._get_list_field_names` using `get_inner_types`
+- switch from poetry to pdm
+- use vocabulary JSON files from mex-model
+
+### Removed
+
+- remove vocabulary JSON files
 
 ### Fixed